risk_engine/requirements.txt
fastapi
uvicorn
pymongo
motor
pydantic
redis
python-dotenv
pytest
email-validator
scikit-learn
joblib
pandas
